Temperature reducer and action
