+++ client/src/redux/actions/temperature.js
@@ -1,0 +1,7 @@
+import { LOAD_TEMPERATURE } from '../reducers/temperature';
+
+export const loadTemperature = () => ({
+  type: LOAD_TEMPERATURE,
+});
+
+export default loadTemperature;
